requirements.txt
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
filelock==3.13.1
fsspec==2024.6.1
google-auth==2.40.2
google-auth-oauthlib==1.2.2
gspread==6.2.1
gspread-dataframe==4.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
oauthlib==3.2.2
pandas==2.2.3
pillow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dateutil==2.9.0.post0
python-telegram-bot==22.1
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==70.2.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.3
torch==2.7.0+cu118
torchaudio==2.7.0+cu118
torchvision==0.22.0+cu118
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
